+++ test/cljs/lambdacd/runner.cljs
@@ -8,7 +8,8 @@
             [lambdacd.time-test]
             [lambdacd.ui-core-test]
             [lambdacd.db-test]
-            [lambdacd.logic-test]))
+            [lambdacd.logic-test]
+            [lambdacd.console-output-processor-test]))
 
 (doo-tests 'lambdacd.history-test
            'lambdacd.output-test
@@ -18,4 +19,5 @@
            'lambdacd.time-test
            'lambdacd.ui-core-test
            'lambdacd.db-test
-           'lambdacd.logic-test)
+           'lambdacd.logic-test
+           'lambdacd.console-output-processor-test)